requirements.txt
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
boto3==1.35.2
botocore==1.35.2
certifi==2024.7.4
cffi==1.17.0
Django==5.1
django-filter==24.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.1
inflection==0.5.1
jmespath==1.0.1
mccabe==0.7.0
minio==7.2.8
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2